completed

+++ app/src/main/java/com/example/final_project/FoodData.kt
@@ -2,15 +2,15 @@ package com.example.final_project
 
 data class FoodData(
     val food_name: String,
-    val nf_calories: Float,
-    val nf_total_fat: Float,
-    val nf_saturated_fat: Float,
-    val nf_cholesterol: Float,
-    val nf_sodium: Float,
-    val nf_total_carbohydrate: Float,
-    val nf_dietary_fiber: Float,
-    val nf_sugars: Float,
-    val nf_protein: Float
+    val nf_calories: Double,
+    val nf_total_fat: Double,
+    val nf_saturated_fat: Double,
+    val nf_cholesterol: Double,
+    val nf_sodium: Double,
+    val nf_total_carbohydrate: Double,
+    val nf_dietary_fiber: Double,
+    val nf_sugars: Double,
+    val nf_protein: Double
 )
 
 data class NutritionResponse(
